Fix jump bug
Default cmc behavior resets CharacterOwner->bPressedJump to false if the JumpMaxHoldTime is surpassed. I want "jumping" to be active as long as you hold it, so I made the JumpMaxHoldTime very large. Had to reset the property in the blueprint character too since that saved the old value of 0 instead of taking the new default set in the c++ constructor.